Preserve native YUV/HDR, complete Hap HDR/R, and enable HEVC inter decode (#200)
* video: decode Hap R / BC7 frames

* tests: cover Hap R / BC7 decode path

* hevc: fix spatial merge candidate availability

* Add HDR and planar YUV pixel formats

* Describe raw image color interpretation

* Teach RawImage about HDR and planar YUV layouts

* Add conversion boundary for extended raw formats

* Route writer coercion through extended raw converter

* Test HDR and planar RawImage layouts

* Add native planar RawImage factory

* Return native VP9 YUV planes

* Make YUV matrix fallback explicit

* Return native H.264 YUV planes

* Return native H.265 YUV planes

* Add exact BC6H half-float decoder

* Decode Hap HDR to native half-float RGB

* Test Hap HDR native half-float output

* Test H.264 native YUV at the consumer boundary

* Test VP9 native YUV at the consumer boundary

* Scope packed pixel conversion matrix to packed formats

* Update PixelFormat cardinality for HDR and planar layouts

* Enable HEVC predicted and bidirectional slice parsing

* Exercise HEVC inter slices beyond slice type gate

* Generalize 8-bit planar YUV image factory

* Generalize VP9 frame buffers for profile 1 chroma

* Preserve VP9 profile chroma geometry in reference frames

* Make VP9 inter prediction subsampling-aware

* Keep VP9 profile 0 call compatibility during profile 1 work

* Parse VP9 profile 1 color and chroma configuration

* Add canonical planar YUV 4:4:0 formats

* Teach RawImage planar helpers about YUV 4:4:0

* Add 8-bit YUV 4:4:0 image factory

* Update pixel format coverage for YUV 4:4:0

* Expose VP9 profile 1 native chroma and sRGB output

* Map VP9 color-space codes without widening constants API

* Make VP9 loop filtering profile-1 subsampling-aware

* Carry VP9 profile subsampling into legacy prediction call sites

* Configure VP9 inter prediction for each frame profile

* Include VP9 color interpretation in frame-pool identity

* Keep VP9 frame-pool color interpretation exact

* Verify VP9 profile 1 against libvpx lossless frames

* Update VP9 boundary tests for profile 1 support

* Parse VP9 profiles 2 and 3 high-bit-depth headers

* Store VP9 reconstruction at native sample precision

* Track VP9 bit depth in reference frame buffers

* Make VP9 intra prediction bit-depth aware

* Make VP9 inter prediction preserve high bit depth

* Extend VP9 loop filter to 10 and 12 bit samples

* Add normative VP9 10 and 12 bit quantiser tables

* Reconstruct VP9 at native 8 10 or 12 bit precision

* Remove VP9 profile-1 prediction compatibility state

* Expose all VP9 profiles at native sample precision

* Add libvpx oracle tests for VP9 profiles 2 and 3

* Retire VP9 high-bit-depth refusal tests

* Parse HEVC tile geometry instead of refusing it

* Add HEVC raster-to-tile CTB address mapping

* Support HEVC dependent slice header inheritance

* Wire HEVC dependent slice inheritance at picture level

* Decode HEVC tiles WPP and dependent segments in normative CTB order

* Honor HEVC tile boundaries in sample adaptive offset

* Honor HEVC loop-filter-across-tiles flag

* Update HEVC boundary tests for tiles and dependent segments

* Add H.264 scaling-list parsing and fallback rules

* Retain H.264 High-profile scaling lists and direct inference flag

* Retain H.264 High-profile PPS scaling and list1 state

* * make the branch build, and update the H.264 test the change intends

_Chroma still took byte[] after the VP9 planes widened to ushort[], so nothing
compiled.

ScalingMatricesAreRefusedByName asserted a refusal this branch deliberately
removes by implementing the matrices. It now asserts the stream decodes.

Eight failures remain and all are this branch's own VP9 high-bit-depth tests.

* * VP9 category-six coefficients are wider than fourteen bits above eight-bit

Category six is the one token whose bitstream depends on sample depth: fourteen
extra bits at eight bits, sixteen at ten, eighteen at twelve, each read from its
own tail of vp9_cat6_prob_high12. This read fourteen with the eight-bit table
whatever the stream said, so the arithmetic decoder fell two bits behind at the
first large coefficient and everything after it was noise.

It looked like a transform or prediction fault — samples sat near the DC
predictor, 512 at ten bits, with the error appearing at an eight-sample
boundary. It was neither. Row 0 of the profile-2 fixture decoded as
785 783 786 787 786 790 783 789 513 510 ... against an expected ramp of
123 140 157 ...; it now decodes as the ramp exactly.

Two fixtures were also invalid base64 — 613 and 407 characters, and 613 is a
length no base64 string can have. Both were regenerated with libvpx-vp9
-lossless 1 from the source formulas the test states, and each round-trips
through ffmpeg byte-for-byte before being embedded, which is what the file's own
remark says its fixtures must do.

Checked past the fixtures as well: a three-frame lossy 64x48 profile-2 stream
decodes byte-identical to ffmpeg across all three planes.

2,592 video tests pass.

+++ Hawkynt.FileFormats.Video/Codecs/Vp9/Vp9ColorConversion.cs
@@ -82,7 +82,7 @@ internal static class Vp9ColorConversion {
   /// distance written as weights. Past the edge of the plane the near sample is used for both, which
   /// replicates the edge rather than wrapping.
   /// </remarks>
-  private static int _Chroma(byte[] plane, int stride, int width, int height, int x, int y) {
+  private static int _Chroma(ushort[] plane, int stride, int width, int height, int x, int y) {
     var nearX = x >> 1;
     var nearY = y >> 1;
     var farX = _Neighbour(nearX, x, width);
